pyproject.toml
[project]
name = "living-docs-sync"
version = "1.0.0"
description = "Living Docs Architecture Synchronizer - MCP-powered Notion ↔ GitHub sync"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "typer>=0.12.0",
    "rich>=13.9.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "anthropic>=0.42.0",
    "openai>=1.58.0",
    "google-generativeai>=0.8.0",
    "unidiff>=0.7.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "python-dotenv>=1.0.0",
    "google-genai"
]

[project.scripts]
living-docs = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*", "api*", "core*", "storage*", "utils*"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.uv]
dev-dependencies = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]
